started reorganizing Types from notebook-common and type converters
