+++ packages/notebook/src/browser/view-model/notebook-model.ts
@@ -20,9 +20,10 @@ import {
     CellData,
     CellEditOperation, CellEditType, CellUri, NotebookCellInternalMetadata,
     NotebookCellsChangeType, NotebookCellTextModelSplice, NotebookData,
-    NotebookDocumentMetadata, NotebookModelWillAddRemoveEvent,
-    NullablePartialNotebookCellInternalMetadata, NotebookContentChangedEvent
+    NotebookDocumentMetadata,
+    NullablePartialNotebookCellInternalMetadata
 } from '../../common';
+import { NotebookContentChangedEvent, NotebookModelWillAddRemoveEvent } from '../notebook-types';
 import { NotebookSerializer } from '../service/notebook-service';
 import { FileService } from '@theia/filesystem/lib/browser/file-service';
 import { NotebookCellModel, NotebookCellModelFactory } from './notebook-cell-model';
