+++ packages/plugin-ext/src/plugin/type-converters.ts
@@ -31,7 +31,7 @@ import { DisposableCollection, Mutable, isEmptyObject, isObject } from '@theia/c
 import * as notebooks from '@theia/notebook/lib/common';
 import { CommandsConverter } from './command-registry';
 import { BinaryBuffer } from '@theia/core/lib/common/buffer';
-import { CellData, CellExecutionUpdateType, CellOutput, CellOutputItem, CellRange, isTextStreamMime } from '@theia/notebook/lib/common';
+import { CellRange, isTextStreamMime } from '@theia/notebook/lib/common';
 import { CellExecuteUpdate, CellExecutionComplete } from '@theia/notebook/lib/browser';
 import { MarkdownString as MarkdownStringDTO } from '@theia/core/lib/common/markdown-rendering';
 
@@ -1634,126 +1634,6 @@ export namespace NotebookKernelSourceAction {
             documentation: item.documentation
         };
     }
-}
-
-export namespace NotebookDto {
-
-    export function toNotebookOutputItemDto(item: CellOutputItem): rpc.NotebookOutputItemDto {
-        return {
-            mime: item.mime,
-            valueBytes: item.data
-        };
-    }
-
-    export function toNotebookOutputDto(output: CellOutput): rpc.NotebookOutputDto {
-        return {
-            outputId: output.outputId,
-            metadata: output.metadata,
-            items: output.outputs.map(toNotebookOutputItemDto)
-        };
-    }
-
-    export function toNotebookCellDataDto(cell: CellData): rpc.NotebookCellDataDto {
-        return {
-            cellKind: cell.cellKind,
-            language: cell.language,
-            source: cell.source,
-            internalMetadata: cell.internalMetadata,
-            metadata: cell.metadata,
-            outputs: cell.outputs.map(toNotebookOutputDto)
-        };
-    }
-
-    // export function toNotebookDataDto(data: NotebookData): rpc.NotebookDataDto {
-    //     return {
-    //         metadata: data.metadata,
-    //         cells: data.cells.map(toNotebookCellDataDto)
-    //     };
-    // }
-
-    export function fromNotebookOutputItemDto(item: rpc.NotebookOutputItemDto): CellOutputItem {
-        return {
-            mime: item.mime,
-            data: item.valueBytes
-        };
-    }
-
-    export function fromNotebookOutputDto(output: rpc.NotebookOutputDto): CellOutput {
-        return {
-            outputId: output.outputId,
-            metadata: output.metadata,
-            outputs: output.items.map(fromNotebookOutputItemDto)
-        };
-    }
-
-    export function fromNotebookCellDataDto(cell: rpc.NotebookCellDataDto): CellData {
-        return {
-            cellKind: cell.cellKind,
-            language: cell.language,
-            source: cell.source,
-            outputs: cell.outputs.map(fromNotebookOutputDto),
-            metadata: cell.metadata,
-            internalMetadata: cell.internalMetadata
-        };
-    }
-
-    // export function fromNotebookDataDto(data: rpc.NotebookDataDto): NotebookData {
-    //     return {
-    //         metadata: data.metadata,
-    //         cells: data.cells.map(fromNotebookCellDataDto)
-    //     };
-    // }
-
-    // export function toNotebookCellDto(cell: Cell): rpc.NotebookCellDto {
-    //     return {
-    //         handle: cell.handle,
-    //         uri: cell.uri,
-    //         source: cell.textBuffer.getLinesContent(),
-    //         eol: cell.textBuffer.getEOL(),
-    //         language: cell.language,
-    //         cellKind: cell.cellKind,
-    //         outputs: cell.outputs.map(toNotebookOutputDto),
-    //         metadata: cell.metadata,
-    //         internalMetadata: cell.internalMetadata,
-    //     };
-    // }
-
-    export function fromCellExecuteUpdateDto(data: rpc.CellExecuteUpdateDto): CellExecuteUpdate {
-        if (data.editType === CellExecutionUpdateType.Output) {
-            return {
-                editType: data.editType,
-                cellHandle: data.cellHandle,
-                append: data.append,
-                outputs: data.outputs.map(fromNotebookOutputDto)
-            };
-        } else if (data.editType === CellExecutionUpdateType.OutputItems) {
-            return {
-                editType: data.editType,
-                append: data.append,
-                outputId: data.outputId,
-                items: data.items.map(fromNotebookOutputItemDto)
-            };
-        } else {
-            return data;
-        }
-    }
-
-    export function fromCellExecuteCompleteDto(data: rpc.CellExecutionCompleteDto): CellExecutionComplete {
-        return data;
-    }
-
-    // export function fromCellEditOperationDto(edit: rpc.CellEditOperationDto): CellEditOperation {
-    //     if (edit.editType === CellEditType.Replace) {
-    //         return {
-    //             editType: edit.editType,
-    //             index: edit.index,
-    //             count: edit.count,
-    //             cells: edit.cells.map(fromNotebookCellDataDto)
-    //         };
-    //     } else {
-    //         return edit;
-    //     }
-    // }
 }
 
 export namespace TestMessage {
